adjusting reconst test and moving reconst method to helpers
